+++ gcs/src/test/java/com/google/cloud/hadoop/fs/gcs/GoogleHadoopFileSystemTest.java
@@ -394,7 +394,7 @@ public class GoogleHadoopFileSystemTest extends GoogleHadoopFileSystemIntegratio
   public void testRenameHnBucket() {}
 
   @Override
-  public void testRenameWithMoveOperation() {}
+  public void testRenameWithMoveDisabled() {}
 
   @Override
   public void testHnBucketRecursiveDeleteOperationOnDirectory() {}